WFCORE-1059 RestartParentResourceAdd/RemoveHandler should handle capability registration
